requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tensorflow-cpu==2.13.0
librosa==0.10.1
numpy==1.24.3
pandas==2.1.3
matplotlib==3.7.2
seaborn==0.13.0
plotly==5.17.0
scikit-learn==1.3.2
psutil==5.9.6
python-multipart==0.0.6
opencv-python-headless==4.8.1.78
python-dotenv==1.0.0
locust==2.16.1
soundfile==0.12.1
resampy==0.4.2
